+++ src/frontend/screens/ManualGpsScreen/DmsForm/types.ts
@@ -1,0 +1,7 @@
+export type DmsData = {
+  degrees: string;
+  minutes: string;
+  seconds: string;
+};
+
+export type DmsUnit = keyof DmsData;
